ready for parsing and exec merging
