+++ src/parsing/parsing.c
@@ -6,7 +6,7 @@
 /*   By: ccolin <ccolin@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/12 13:35:31 by ccolin            #+#    #+#             */
-/*   Updated: 2024/11/17 11:48:16 by ccolin           ###   ########.fr       */
+/*   Updated: 2024/11/17 13:24:14 by ccolin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
